+++ Source_Files/Input/mouse_sdl.cpp
@@ -101,7 +101,7 @@ void mouse_idle(short type)
 {
 	if (mouse_active) {
 		static uint32 last_tick_count = 0;
-		uint64 tick_count = machine_tick_count();
+		uint32 tick_count = machine_tick_count();
 		int32 ticks_elapsed = tick_count - last_tick_count;
 
 		if (ticks_elapsed < 1)
